This check-in provides a work-around to the error message you sometimes get about being unable to retrieve tickmark lengths:
  warning:tmYLMinorLengthF is not a valid resource in
          gsn_csm_contour_map.PlotManager at this time

This error message can occur if you're using 2D lat/lon coordinates,
and it goes into a mode where tickmarks are turned off. This causes a
problem for some internal code that tries to retrieve tickmark lengths
so it can make them the same size on all axes (this was
pre-tmEqualizeSizes).

This code will now be skipped if res@gsnScale is set to False.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9948 10fe2890-8a40-4696-8b1d-7ccbb19c398e