Corrected an error message in Grib2ReadCodeTables: it was prefixing the full path name with the directory giving a misleading message.
Also fixed a core dump that was occuring if the centers.table file was not
found. Now this is considered a fatal error, because it implies that the
GRIB2 codetable directory path is not set correctly.

Also modified the way the codetable path is determined. Now the
NIO_GRIB2_CODETABLES environment variable is checked first since it is
more specific and therefore should take precedence. This will allow a
user to have their own codetable directory override a common one found under
NCARG_ROOT.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9955 10fe2890-8a40-4696-8b1d-7ccbb19c398e