Updated "conform" to accept strings as the type for the second argument.
Added a new "conform_dims" function which behaves identically to
"conform", except instead of passing an array that you want to conform
a smaller array to, you just pass the dimension sizes of the
array you want to conform to. For example:

Using "conform" to conform a 1D array to a 3D array:

    x1d = (/8.,9.,10./)
    x3d = new (/(/2,3,4/),typeof(x1d))
    x3d = conform (x3d, x1d, 1)

Using "conform_dims" to conform the same 1D array to a 3D array:

    x1d = (/8.,9.,10./)
    x3d = conform_dims ((/2,3,4/), x1d, 1)   ; x3d will be 2 x 3 x 4

Note this second method is preferable in most cases, because
you don't need to create "x3d" apriori.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9991 10fe2890-8a40-4696-8b1d-7ccbb19c398e