This check in changes the GRIB 1 reader to treat grid number 0 as a synonym of grid number 255 (i.e. grid is not pre-defined, use the GDS part of the record to determine the grid).
Also changes some instances of the Dx and Dy attribute values to use meters
rather than kilometers as units. This is for consistency with the majority of
the instances.

Also changed the llmap.f code that contains a few EZMAP routines that are
needed for generating GRIB coordinate values when libncarg is not linked in (i.e.
when compiling libnio for PyNIO). The changes are the same as Dave Kennison
is making to correct the problem involving setting up a Lambert Conformal projection
with certain 'difficult' values used for the standard parallels.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10006 10fe2890-8a40-4696-8b1d-7ccbb19c398e