This code checks in *some* of the changes needed to remove the internal png and zlib code from the GKS library.  There are a few other changes (in two "scripts" directory and the "hlu" library directory) I still need to check in.
In pixcontrol.c, an "#ifdef BuildPNG" and "#endif" block was inserted
around the "case PIX_PNG" code, to prevent the user from being able to
open a PNG image workstation (if PNG support was not indicated as
being desired).

The BuildPNG macro gets set if the user runs "./Configure" and answers
"yes" to the question about whether PNG support is desired.

In pixwrite_xwd.c, there were lots of "printf" statements that looked
like debug statements, so I commented them out. I've left a note for
Dave to check this.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10019 10fe2890-8a40-4696-8b1d-7ccbb19c398e