Support added for the setfileoption choice TimePeriodSuffix, described by Dave as:
This checkin implements a new file option for GRIB 1 and 2 files known
as "TimePeriodSuffix". It is a logical option with default value True.
If set False the suffix representing a time period (e.g. "2h") are removed
from statistically processed variables, leaving only the type of processing
as a suffix (e.g. "_acc", "_avg"). This can help when the NIO library
gives different names to variables (ususally from different files)
that the user wants to concatenate together typically using addfiles.

Note that when addfiles_GetVar is used with GRIB 2 files (as in Dennis' test
script) the "statistical_processing_duration" attribute may no longer
be meaningful. It would be more correct in this case to use the value of the
attribute given by the third file in the list rather than the first one,
but I see no way to correct for that problem.

Specifying "-tps" on the ncl_convert2nc command line will change the logical
value of TimePeriodSuffix to False.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10040 10fe2890-8a40-4696-8b1d-7ccbb19c398e