This is similar to the change Dave K made to $NCARG/ncarg2d/src/liboptional/ncarbd/ncarbd.f.
This routine gets loaded when you specific "-ngmathbd" on the
ncargf77, ncargcc, nhlcc, or nhlf77 compile line. It's been updated to
reflect the change in names of the two blockdata statements in the
Ngmath library.

Since blockdata initialization problems are hopefully a thing of the
past, this change probably shouldn't matter to anybody.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10121 10fe2890-8a40-4696-8b1d-7ccbb19c398e