Fixed bug found by Dave Brown when he tried to run Liu Yan's WRF-NCL script.
The fix was simple: an input argument to a function (opt_args) was
being referenced inside a routine and possibly changed. (It didn't
appear to be getting changed, but I think NCL thought it was.) This
caused NCL to issue a syntax error message somewhere later in the
code, in a place that had nothing to do with "opt_args". This was a
bit confusing.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10124 10fe2890-8a40-4696-8b1d-7ccbb19c398e