Changes to make Earth..4 fill work for MapPlot.
The only missing item is a resource to control the outlines of the Antarctic ice shelf.
I increased the default AreaGroupCount by 1 in order to provide a group for the ice shelf.
This means the ice shelf fill color is controllable: it has its own color by default, but
you can easily color it the same as the continent (or the ocean) if you want.
The states (provinces) of China, Brazil, Australia, and India can be accessed just
like those of Canada and Mexico.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10344 10fe2890-8a40-4696-8b1d-7ccbb19c398e