(In some of these routines, only the "$Id$" line has changed, but there are enough others to make it difficult to sort out only the ones that really need to be checked in.  This is due mostly to the fact that I am using "Fsplit" to extract the routines from a monolithic file; it supplies each with a $Id$ line and that's enough for CVS to consider it changed.)
The most important effect of this update is to simplify the code for both
the forward and inverse forms of the cylindrical and mixed projections.
The code we have been using was based on the azimuthal projections and
was strangely (and unnecessarily) complicated: it spent a lot of time
doing something necessary for the azimuthals and then undoing it so as
to generate values for the others.  (It was almost as if one wanted to
fly from Denver, Colorado, to Washington, D. C., and chose to go by way
of Saint Paul, Minnesota, and some dinky little town in Alaska ...:-)
I have suspected for some time that there was a better way; putting in
the new projections led me to work it out.

Along the way, I also fixed some variable names.  At some time in the past,
someone used names evoking "lambda" and "phi" for latitude and longitude,
respectively, and that is exactly backwards: in all the reference books and
papers, "lambda" is used for longitude and "phi" for latitude.  Again, I
have known about this for some time, but was reluctant to change the code
for fear of screwing it up.  At this point, however, the reversed names are
unduly increasing the mental gymnastics required to understand the code, so
I went ahead and changed them (and, in fact, I mostly stuck with names that
suggest "latitude" and "longitude", thereby skirting the issue).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10419 10fe2890-8a40-4696-8b1d-7ccbb19c398e