Modified ncl_convert2nc to restore the -itime and -ftime options (for better backwards compatiblity and convenience). Rick had removed these when he added the -sed option. Now these options can be used together or separately. It doesn't matter if a dimension type is specified more than once.
Modified ncl_filedump to support the -sed option. It still supports -itime and
-ftime. All these options work alike for both scripts now.
Removed an extraneous print statement from the output when the '-c' option
is used. Also changed the sub-heading that precedes coordinate data output
from 'data:' to 'Coordinates:'. This seems to match the sub-heading for
variable output ('Variables") better.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10441 10fe2890-8a40-4696-8b1d-7ccbb19c398e