Fred reported that this example (from the contouring and mapping tutorial) was not working properly for him, so I took a look at it and found various problems, both in the example and in the part of the tutorial describing it. I have made various changes aimed at making the comments in the code more nearly consistent with the code itself and in making the code more usable.
Fred reported one behavior that I was unable to duplicate: the apparent
disappearance of some lines, due to their having been drawn with the same
color index as the background.  I do not know if I have fixed this, but no
doubt Fred will let me know if I have not.

I will make some changes in the text of the tutorial, as well.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10462 10fe2890-8a40-4696-8b1d-7ccbb19c398e