Added: wgt_runave_leftdim and dtrend_leftdim
These are utility functions that result in cleaner code.

The running average or detrending is done on the
leftmost dimension [hence, the appended 'leftdim']
Generally, this is the 'time' dimension.


Internally,

(a) arrays is reordered
(b) the wgt_runave_Wrap or dtrend is applied
(c) the arrays are reordered and returned in the original order

Doc will follow at sometime before the release of 5.0.1


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10475 10fe2890-8a40-4696-8b1d-7ccbb19c398e