Mary,
This requires *no* changes in the previous interface:
     'mjo_cross_segment'

However, I did break out the smoothing  section into a separate
subroutine. This is called by the driver. However, another
application needs the same smoother.

I need call the smoothing from NCL script.

This will be a procedure. The "STC" returned by 'mjo_cross_segment'
is input to the procedure .... The 1:8 entries are used
to alter the 9:16.

       mjo_cross_coh2pha (STC)   ; alter the coherence^2 & phase

D


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10543 10fe2890-8a40-4696-8b1d-7ccbb19c398e