This check-in is a stage in the implementation of an independently-buildable libnio.a. Basically it adds a preprocessor define called NIO_LIB_ONLY that is used to selectively include code based on whether the files are compiled within the NCL/NCARG context or whether they are being independently built for libnio.a. A typical usage is:
#ifdef NIO_LIB_ONLY
#include "niohlu.h"
#include "nioNresDB.h"
#include "nioCallbacks.h"
#else
#include <ncarg/hlu/hlu.h>
#include <ncarg/hlu/NresDB.h>
#include <ncarg/hlu/Callbacks.h>
#endif

The Nio library, when compiled independently, has its own versions of HLU
and NCARG header files that have been stripped of references to the NCARG
and GKS libraries. These are all headers files that are pretty stable and
will not need to change in the NIO context. As long as NIO_LIB_ONLY is False
there should be no change in the behavior of NCL.

Additionally the yMakefile has been modified to copy the source
files that are common to NCL and libnio.a into a <>/ncarg/pynio subdirectory
called libsrc during the make install phase. Eventually the files currently
residing in the NCL source directory that are only used for libnio.a will
be removed. They will live permanently in <>/ncarg/pynio/libsrc.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10575 10fe2890-8a40-4696-8b1d-7ccbb19c398e