Adds "Earth..1", "Earth..2", "Earth..3", and "Earth..4" to the list of database files that are installed.  This is a first step in the direction of making the Ezmap database editor available.
I still haven't completely decided just how public to make the editor, but
I'm leaning strongly in the direction of creating a programmer document for
it and making it downloadable from there.

In order for the editor to be fully functional, one has to download the
RANGS/GSHHS files, put them somewhere, and then modify the code of the
editor to tell it where they are.  Practically speaking, then, it makes
no sense to install it along with everything else.  Add to that the fact
that one of the strengths of the editor is that one can modify it to do
specific things and it makes more sense to give users the code and tell
them how to compile and run it from their own directories.

I suppose the files "Earth..n" could also be made downloadable from the
programmer document, but they are already in the installation directories;
they just haven't been installed heretofore.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10587 10fe2890-8a40-4696-8b1d-7ccbb19c398e