This burned up a number of hours, but it turns out to be a relatively minor change ...
The image of Antarctica that I used to create new coastlines and ice-shelf
boundaries was originally billed, on the site from which I downloaded it, as
a satellite-view image from the distance of the moon, so I georeferenced and
used it as such.  As I began to document the process of importing and using
image files in the database editor, I did some playing around with a global
image that ought to give identical results for Antarctica and observed a
slight discrepancy between its boundaries and mine (less than a tenth of a
degree, but certainly enough to notice).  I went back to the site from which
I got the Antarctica image and it is now billed as an orthographic view -
almost the same, but not quite.  Luckily, it was relatively easy, once I
wrapped my aging brain around the math, to program the necessary slight
shift in all the boundaries that I digitized (instead of redigitizing them
all, which would have been an impossible task in the time I have left).

No area identifiers were changed in the process and I see no reason to redo
examples that are based on "Earth..4" (I know that Dave Brown did one or more
specifically for Antarctica), as the visual effects of my change should be
negligible.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10661 10fe2890-8a40-4696-8b1d-7ccbb19c398e