Added support for NCEP parameter table 132, the new number for the reanalysis table.
Fixed a bug in the NetCDF I/O module that resulted in an incorrect value
for character variable missing values. This was a result of NCL converting all
character valued attributes to strings for easier manipulation in NCL.
This does not work for the _FillValue of character type variables, because
of the fact that NCL insists (for good reason) that the _FillValue have the
same type as the variable it belongs to.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10691 10fe2890-8a40-4696-8b1d-7ccbb19c398e