This fixes a bug reported by Xuguang Sun where upon assigment to a previously defined variable or to a file variable all the elements of a value assigned from a filelist variable take the value of the first element.
This problem affects "cat" mode only and is specific to lists of files where
the aggregated (leftmost) dimension of a variable contains a single element in
the first file in the list. In this situation the multi-dvals that contain
the data for the variable and the coordinate variable has the "kind" structure
member set to SCALAR (because at this point it is SCALAR). However, the list
filevar code modifies this variable in place, changing it to a non-scalar
multidval. However, it was failing to modify the "kind" member to MULTID.
This is now corrected.

Note that when the variable data has multiple dimensions the data multidval
will have its kind element set to MULTID, but the leftmost coordinate dimension (in
the situation above) still has only one element and thus is "kind" SCALAR.

Having this flag set incorrectly leads in some cases to scalar to array assignment.
In other words the first element of the array gets assigned to all elements of
the target variable array. This does not happen in all cases, and the behavior
seems quite peculiar, but it is explainable.

I will send an email to ncl-talk explaining a work-around for this problem.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10731 10fe2890-8a40-4696-8b1d-7ccbb19c398e