Fixes a Fortran error that fort77 on a Linux machine caught in stutil.f.
Increases the 'crossover check' circular list length by more than half. This
eliminates an instance where points that still needed to be checked were being
discarded because the list was not long enough. The result was an infinite loop
caused by a streamline that followed the same path over and over.
Also introduced a new check designed to prevent an infinite loop if the list
again proves to be too short. It stops the streamline generation if the number
of iterations gets way over the norm (at least I hope there are not instances
where a legitimate streamline would become as big as the size specified.) It
is set dynamically based on the step size in NDC as the number of step required
to create a line that crosses the whole viewspace 20 times.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10787 10fe2890-8a40-4696-8b1d-7ccbb19c398e