This change does the same thing Dave just did for BuiltInFuncs.c:
    This changes all cases of calling _NclGetArg or _NclGetArgValue
    with the final parameter set to 2 (READ_IT) to the value 0
    (DONT_CARE). These defines dont actually mean what they
    say. Setting 0 (DONT_CARE) actually results in read-only access to
    the argument; whereas 2 allows it to change.

I'm a bit nervous about this change, because there may be some cases
where a variable is supposed to be an input/output variable,  and
if it (incorrectly) had a "2" as the last argument to NclGetArgValue,
then it now has a value of "0", which will prevent it from being
an output value.

All the current tests ran correctly, so we'll just have to hope that
we don't have any untested functions that have both input/output
variables.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10791 10fe2890-8a40-4696-8b1d-7ccbb19c398e