Another stab at fixing the build macros for cairo, gdal, and everybody else for that matter.
If you want cairo and/or gdal support in your software, add the
following lines to $NCARG/config/Site.local:

#define BuildGDAL 1
#define BuildCAIRO 1

This stuff is not built by default, so if you don't want one or the
other, then make sure there is nothing referencing cairo or gdal
in your Site.local file.

----------------------------------------------------------------------
A more detailed description:

The gist of the problem was that Configure was setting macros with:

#define BuildGDAL FALSE

and then the Project file was trying to evaluate these with:

#if !BuildGDAL

BuildGDAL will only define to FALSE if it is not set, or if
it is set to 0.

So, I cleaned this up, and all macros written by Configure now use 1/0
instead of TRUE/FALSE.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10837 10fe2890-8a40-4696-8b1d-7ccbb19c398e