With respect to the bug Mark Branson reported in vinth2p:
   I was taking a close look at the vinth2p routine recently and I
   found something odd about how it handles the vertical interpolation
   for the very top output pressure levels.

He and Dennis determined that KP should be set to 0, and not 1,
in one place in the code.  Note that there's still another
place in the code where it is being set to 1, but I'm not going
to change this until Dennis does more research.

This is true for all three

Secondly, we also became aware that the PLEVI array is dimensioned
NLEVI, but the calling C program is allocating NLEVI+1 for this,
*and* the fortran code seems to indicate that PLEVI could
have an NLEVI+1-th value, so I've changed this as well.

The calling program will need to be fixed because it wasn't
passing nlevi+1 correctly.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10900 10fe2890-8a40-4696-8b1d-7ccbb19c398e