This is a fix for the bug involving conversion of integer to logical data exemplified by this code:
allw            = (/1,2,3,4,-999,-999,5,6,-999/)
allw@_FillValue = -999
print(ind(allw))

It correctly produces the following output now:

warning:Argument 0 of the current function or procedure was coerced to the appropriate type and thus will not change if the function or procedure modifies its value
(0)     0
(1)     1
(2)     2
(3)     3
(4)     6
(5)     7

The problem was introduced by my changes to allow the logical missing value to be set to e.g. False rather than _Missing. The new code allows the
missing value to be set, but prevents it from being modified when other types such as integer are coerced to logical. Otherwise, a missing integer or
float value of -999 would cause the logical missing value to change to True, which is in my opinion rather undesireable.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10922 10fe2890-8a40-4696-8b1d-7ccbb19c398e