Fix for PYNIO-4 (it affects NCL equally but I guess there is no NCL ticket) Removed the following from the yMakefile: # HDF4.2r2 requires this definition because we ask users to # build it w/o internal netCDF support HDF4_NETCDF = -DH4_HAVE_NETCDF
With previous versions of the HDF library the define 'HAVE_NETCDF' meant that an
external NetCDF library was being supplied, and therefore to build NCL it had to
be defined. With newer versions of HDF the define H4_HAVE_NETCDF has an opposite
meaning: that NetCDF is contained within the HDF library. How confusing!

Anyway we need not to define this preprocessor variable.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@10971 10fe2890-8a40-4696-8b1d-7ccbb19c398e