Fix for NCL-281.
When assignment is made using subscripting, the coordinate variables get an extra parent
reference. The 'delete' function was stepping through the references and removing them
one at a time, but to do that for coordinate variables it calls _NclDeleteCoordVar which
internally deletes the whole reference list. This meant the 'delete' routine ended up
trying to dereference a NULL pointer. Now it breaks immediately from the loop after
the call to _NclDeleteCoordVar.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11003 10fe2890-8a40-4696-8b1d-7ccbb19c398e