This hopefully provides a "fix" for NCL-286.
I put "fix" in quotes because I'm not changing the behavior of the
gsn_csm_xxxx routines. Rather, the gsn_add_polyxxxx routines now
recognize the "tfPolyDrawOrder" resource.

This allows you to control the draw order of each primitive that you
add to a plot ("PreDraw", "Draw", or "PostDraw").

For example, in the code below, we are attaching a polygon to
both the contour plot and the map plot. In the contour plot,
the polygon will be drawn in the last (postdraw) phase, and in the
map plot it will be drawn in the middle (draw) phase:

  res = True
  res@gsnDraw  = False
  res@gsnFrame = False
  ...
  plot = gsn_csm_contour_map(wks,data,res)

  gres = True
  gres@gsFillColor     = "green"
  gres@tfPolyDrawOrder = "Draw"    ; This is new, couldn't do this before.

  dum1 = gsn_add_polygon(wks,plot,lon1,lat1,gres)  ; attach to map

  gres@gsFillColor     = "red"
  gres@tfPolyDrawOrder = "PostDraw"
  dum2 = gsn_add_polygon(wks,plot@contour,lon2,lat2,gres)  ; attach to contour

  draw(plot)
  frame(wks)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11017 10fe2890-8a40-4696-8b1d-7ccbb19c398e