This is a fix for NCL-350 (If _FillValue has type different from variable in file, it is not correctly converted when variable is copied to a local NCL variable).
Now if the file has a variable with a  _FillValue of different type, it is converted
to the correct type when copied to a local NCL variable. This conversion happens
regardless of NCL's usual conversion rules and there is no check for overflow or
loss of precision. However, a warning similar to the following accompanies the variable copy:

warning:FileReadVar: _FillValue attribute type differs from variable (sie) type in file (sie_1973-2001), forcing type conversion; may result in overflow and/or loss of precision

I added a new routine called _NclScalarForcedCoerce. It is like _NclScalarCoerce excepti
that it does all conversions (except from from string and logical -- which do not
appear in files -- yet at least.)


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11097 10fe2890-8a40-4696-8b1d-7ccbb19c398e