This is a fix for NCL-357. It involved rewriting the way GRIB1 Gaussian grid parameters are processed. Actually, I think one could say the GRIB file was incorrect in that the GRIB La1 and La2 values were outside the range of the lat coordinates generated by the ncep gaussian lat code. Nevertheless, it is possible to return valid values (at least I think they are valid). So now there is a warning given for this situation: warning:GdsGAGrid: GRIB attributes La1 and/or La2 are incorrectly out of range of the gaussian latitude array (See GRIB Section 2 documentation) but reasonable lat values are returned.
Fixing this without causing problems with other problematic Gaussian grids required an overhaul of
all the code that determines the Gaussian lat coordinates. I have verified the fix with all GRIB 1 files
that I could find that contain Gaussian grids. It is possible that a similar situation could be encountered
in GRIB 2 files, but I would like to make sure this fix doesn't cause any problems before trying to
change the GRIB2 code, which is necessarily somewhat different.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11119 10fe2890-8a40-4696-8b1d-7ccbb19c398e