This is a fix for NCL-380 and NCL-381.
The code now checks to see that NCL object ids are valid before accepting a
lookup from the NCL HLU object id list. The HLU list still runs into situations
where objects remain in the list after they have been deleted. This will need another
look at some time in the future. But it is a complicated system that imperfectly
relies on callbacks to perform the deletion of objects from the table, when they
are actually deleted deep in the guts of the HLU library. But I think I have put enough
safety checks around the NCL HLU object list that there is no longer a danger of a
reused HLU id mapping into an NCL object id that represents something else entirely (this
was the immediate cause of Jonathan Vigh's core dump).

The fix for NCL-381 is to now require that file objects be created with a _FillValue if
not given an actual valid file object at initialization. Graphic objects also require
initialization but in this case they do not need a fill value. Initializating  the memory
to a negative value is suffficient to ensure that the objects will not be confused with a
valid value.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11165 10fe2890-8a40-4696-8b1d-7ccbb19c398e