This check-in fixes NCL-438. The NCEP code to generate polar stereographic coordinates seems to have trouble with Southern Hemisphere data. I was only using it because it automatically generated the rotation angles in addition to the lat/lon values. Subsequently a rotation function for polar stereograaphic grids was implemented to deal with pre-defined grids where the resolution was too high for the NCEP code to handle. So I reverted to using EZMAP code + the rotation function, which seems to give the correct results for the Southern hemisphere.
Also a small change to static variable initialization for NclNetCdf.c and
I am now treating NCAR center (center 60) data the same as NCEP for the
purposes of parameter definition. As far as I know NCAR has no tables of
its own. This makes all the parameters in the subject file known.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11238 10fe2890-8a40-4696-8b1d-7ccbb19c398e