At Dennis' request, made the following change:
----------------------------------------------------------------------
The current relhum function uses relhum_dp.f
The code is nearly all from the old CCM Processor.

Why "nearly" .. well, after checking, I apparently
added the following line


IF (DRELHUM.GT.100.D0) DRELHUM = 100.D0

Dave Williamson commented to me that this prevents
instances where "super saturation" can occur.

--
Can you please comment that line out?
or,
I can comment and check-in the code?

--
Then it would be in the test 5.2.0 release
for Dave to use.

--
backward compatibility: Super saturation
is a 'corner' case, it should not be a general issue.

I can change the doc to note that the following w\can
be used to force rh to be 0-to-100

    rh = relhum(...)
    rh = where(rh.gt.100, 100, rh)

----------------------------------------------------------------------


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11254 10fe2890-8a40-4696-8b1d-7ccbb19c398e