Dave Brown made changes to this library (Sep 28, 2009) so it could read a MET9_IR108_cosmode_0909210000.grib2 satellite file.
See Jira ticket #NCL-293. Here's a comment from Dave:

I have modified version 1.1.9 of the g2clib to allow NCL to read this
file. Changes were required in 3 files. I also had to make a small
change to the NCL GRIB2 reader. Here is the check-in message for
NclGRIB2.c (version 1.59):

Modified Files:
NclGRIB2.c
Log Message:

This makes NCL more flexible with respect to GRIB2 rotated lat/lon grids.

It seems that the documentation leaves in doubt the question of
whether the lat and lon of the rotated south pole should be adjusted
by the scale factor. Prior to the check in NCL did not apply a scale
factor, and I am thinking that must have worked for some GRIB2
data. Now it checks the absolute value of the number to see whether it
makes sense without applying the scale factor.

The changes to gridtemplates.h do not directly apply to the lack of
recognition of PDS template 4.31 but it turns out that ECMWF and NCEP
have a different idea of whether longitudes can be negative. The
g2clib thinks all longitude values should be positive, and gives weird
results if a longitude is negative. ECMWF, according to the templates
at

http://nwmstest.ecmwf.int/publications/manuals/d/gribapi/fm92/grib2//show/templates

allows negative longitudes (although not everywhere). I think this is
because it is quite inconvenient to always have to use positive
longitudes when you live near the Greenwich Meredian. I should note
that the WMO GRIB2 specification document (at least the versions I
have read) state that longitudes should always be expressed as
positive numbers.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11295 10fe2890-8a40-4696-8b1d-7ccbb19c398e