Fix NCL-30.
Here are the revisions to the GRIB2 code tables. They are based mainly on the
HTML documentation online at http://www.nco.ncep.noaa.gov/pmb/docs/grib2/grib2_doc.shtml.
However, I modified the long names such that only the first word is capitalized and
the rest is lower-case except for acronyms (e.g. ICAP). This is more in line
with the ECMWF documentation. The NCEP docs are somewhat inconsistent in
their capitalization scheme, so this regularizes that aspect of things.

The main concern for backwards compatibility is that NCEP has now given
short names to all the parameters. Formerly, they had not, and so Rick Grubin invented
names where necessary (most of the time). I will try to document at least these
changes in the release notes.
Of course, the names will also change where there was no entry in the table
formerly and the names were assigned as VAR_....

A few units have changed as well, but these are pretty minor.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11349 10fe2890-8a40-4696-8b1d-7ccbb19c398e