This adds some code that I created awhile back to write debug information to a NetCDF file that contains all the resources set by internal WRF scripts. This will help me debug future WRF plotting codes.
This is not complete, however, as I would like to have some code
that automatically writes the debug scripts for you.

Also changed four occurrences of following lines:

     font_height = get_res_value_keep(opt_arg,"FontHeightF",0.01)

to:

     font_height = get_res_value_keep(opts,"FontHeightF",0.01)

"opt_arg" is the original variable passed in, and "opts" is the
copy. It's better to operate on the copy, as the original may just
be a value, like "True" or "False".


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11368 10fe2890-8a40-4696-8b1d-7ccbb19c398e