This is hopefully a fix for NCL-288
   ut_inv_calendar returns incorrect value when calendar type is
   "360" and units are "months since ..."

This bug is a bit nasty because it uses Udunits under the hood to do
some date conversions; these conversions are not aware of the special
360 calendar.  David Pierce, the author of the underlying code for the
360-day calendar, says this Udunits' problem.

For reasons that I don't understand yet, it appears that units of
"years since" and "months since" do not work with the 360 calendar,
but the rest ("days since", "hours since", etc) do work.

For this reason, I decided to take Dave Allured's suggestion of
converting any "years since" or "months since" units to "days since",
doing the calculation as "days since", and then converting back to
"years since" or "months since" values before returning to the user.

This bug fix shouldn't affect any other types of calendars, including
the standard one.

I've added a bunch of tests to see if this works, but I will bug Dennis
for some real-world tests.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11394 10fe2890-8a40-4696-8b1d-7ccbb19c398e