This fixes NCL-600.
Fixed a bug where the Fortran code for dtrend_msg was incorrectly
checking XDT and YDT for being equal to missing. A user
ran into a bug where a missing value that was used, -999, became
an actual value in a calculation for XDT, causing XDT to incorrectly
be considered as missing.

By instead checking the original X and Y for missing, you don't need to
worry about XDT and YDT having calculated values that might be equal
to missing.

Here are the code diffs:

245c245
<          IF (X(N).NE.XMSG .AND. YDT(N).NE.YMSG) THEN
---
>          IF (X(N).NE.XMSG .AND. Y(N).NE.YMSG) THEN
259c259
<          IF (XDT(N).NE.XMSG .AND. YDT(N).NE.YMSG) THEN
---
>          IF (X(N).NE.XMSG .AND. Y(N).NE.YMSG) THEN
282c282
<           IF (XDT(N).NE.XMSG .AND. YDT(N).NE.YMSG) THEN
---
>           IF (X(N).NE.XMSG .AND. Y(N).NE.YMSG) THEN


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11467 10fe2890-8a40-4696-8b1d-7ccbb19c398e