Fix for NCL-608.
This bug was introduced by my fix for NCL-281. There the problem was that
the whole ref-list for the coordinate variable was deleted, causing the
ref-list next pointer to point to freed memory. But I had wrongly assummed
that it always happened this way. In this case, only one element of the list
is deleted per call to the DeleteCoordinate variable function. So the
solution is not to reinitialize the ref list from the coord var object
each time through the loop. I will also be checking this into the 5.2.1 branch.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11496 10fe2890-8a40-4696-8b1d-7ccbb19c398e