Fix for NCL-652, restrictions on graphic output file name length.
The GSN routines use the basename of the output file name as the name of various graphics objects. 
Internally the graphics routines apply suffixes and prefixes (in the case of the sysresfiles) to
this name for various objects and resources. The basic length allowed for these names was only
128 compared to 256 for file names in the LLU. Some routines actually had hard-coded numerical
buffer sizes that were even less (80 characters in the case of XyPlot). This fix makes the length
for resource and object names consistent throughout the HLU code at 320 characters, allowing 
for a 256 character file name plus 64 characters for suffixes. In actuallity only 254 characters
are allowed now for a graphics output filename including the output file type suffix. 





git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_5_2_1@11567 10fe2890-8a40-4696-8b1d-7ccbb19c398e