Fix for NCL-630.
This bug was introduced with the fix for 270, which addressed a problem of
local variables in func/procs not being undefined after the first invocation
of the routine. The fix involved undefining all the symbols in the local 
symbol table for the routine when cleaning up after the routine returns.

However, this table also contains the symbols that represent the parameters
for the routine and they were being set to UNDEF as well. In most cases
this is not a problem except in certain instances, exemplified by the
following code:

color = get_res_value(s,"FontColor1",get_res_value(s,"FontColor2",1))

The same parameter 's' is passed to both the outer and the inner nested
function. When it is set to UNDEF upon completion of the inner nested 
function it becomes undefined for the outer function as well.

The fix is to check each symbol against the symbols used for parameters.
Parameter symbols are no longer set to UNDEF -- only truly local symbols.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_5_2_1@11586 10fe2890-8a40-4696-8b1d-7ccbb19c398e